Add make gen-registry and update gen-contracts to master
