+++ contracts/internal/gen-contracts.go
