+++ go.mod
@@ -8,7 +8,7 @@ require (
 )
 
 // DO NOT CHANGE DIRECTORY: Create a symlink so this works
-// replace github.com/ethereum/go-ethereum => ../celo-blockchain
+replace github.com/ethereum/go-ethereum => ../blockchain
 
 // Use this to use external build
-replace github.com/ethereum/go-ethereum => github.com/celo-org/celo-blockchain v0.0.0-20200519153823-adbdc7f8c27e
+// replace github.com/ethereum/go-ethereum => github.com/celo-org/celo-blockchain v0.0.0-20200519153823-adbdc7f8c27e
